Update minimum commit SHA (baseline) from which dxc was added
This fixes the following error, when running : \dxc-hlsl-2-spirv-test>vcpkg install --feature-flags=registries,manifests
"
The dependency was not found in the baseline, indicating that the package did not exist at that time. This may be fixed by providing an explicit override version via the "overrides" field or by updating the baseline.
See `vcpkg help versioning` for more information.
"